only build an arch if we've got bins for it
git-svn-id: file:///home/svn/framework3/trunk@9316 4d416f70-5f16-0410-b530-b9f4589650da

+++ external/installer-linux/build.sh
@@ -17,27 +17,30 @@ PATH32="framework-${VERSION}-linux-i686.run"
 NAME64="Metasploit Framework v${VERSION} Installer (64-bit)"
 PATH64="framework-${VERSION}-linux-x86_64.run"
 
-rm -rf tmp32
-mkdir tmp32
-cp tmp/msf3.tar tmp32/
-cp bin/linux32.tar.bz2 tmp32/metasploit.tar.bz2
-bunzip2 tmp32/metasploit.tar.bz2
-cp -a scripts/*.sh tmp32/
-cp -a scripts/msfupdate tmp32/
-TMP32=tmp32`date +%s1`
-mv tmp32 $TMP32
-makeself $TMP32 ${PATH32} "${NAME32}" ./installer.sh 32
-rm -rf $TMP32
+if [ -f "${PATH32}" ]; then
+    rm -rf tmp32
+    mkdir tmp32
+    cp tmp/msf3.tar tmp32/
+    cp bin/linux32.tar.bz2 tmp32/metasploit.tar.bz2
+    bunzip2 tmp32/metasploit.tar.bz2
+    cp -a scripts/*.sh tmp32/
+    cp -a scripts/msfupdate tmp32/
+    TMP32=tmp32`date +%s1`
+    mv tmp32 $TMP32
+    makeself $TMP32 ${PATH32} "${NAME32}" ./installer.sh 32
+    rm -rf $TMP32
+fi
 
-rm -rf tmp64
-mkdir tmp64
-cp tmp/msf3.tar tmp64/
-cp bin/linux64.tar.bz2 tmp64/metasploit.tar.bz2
-bunzip2 tmp64/metasploit.tar.bz2
-cp -a scripts/*.sh tmp64/
-cp -a scripts/msfupdate tmp64/
-TMP64=tmp32`date +%s1`
-mv tmp64 $TMP64
-makeself $TMP64 ${PATH64} "${NAME64}" ./installer.sh 64
-rm -rf $TMP64
-
+if [ -f "${PATH64}" ]; then
+    rm -rf tmp64
+    mkdir tmp64
+    cp tmp/msf3.tar tmp64/
+    cp bin/linux64.tar.bz2 tmp64/metasploit.tar.bz2
+    bunzip2 tmp64/metasploit.tar.bz2
+    cp -a scripts/*.sh tmp64/
+    cp -a scripts/msfupdate tmp64/
+    TMP64=tmp32`date +%s1`
+    mv tmp64 $TMP64
+    makeself $TMP64 ${PATH64} "${NAME64}" ./installer.sh 64
+    rm -rf $TMP64
+fi
